Prefer email Date header for INTERNALDATE
When generating an IMAP INTERNALDATE, prefer the Date header from the
email rather than the rtime param passed to the savemessage() function.

The Date header is better information since offlineima's Maildir folder
implementation uses the filesystem mtime as the rtime param, making the
INTERNALDATE the time of last sync not the time of the message delivery.

This particularly matters for Gmail, since it appears to overwrite the
email message's Date header with the IMAP INTERNALDATE, meaning that
syncs from Maildir->Gmail cause all email dates to become the sync dates
rather than the original message dates.